Make engineering mode available for the staging configuration as well
Praise the lord.
