+++ app/build.gradle
@@ -72,8 +72,9 @@ android {
             buildConfigField "String", "WEB_BASE_URL", "\"https://www.dev.hedvigit.com/\""
             resValue "string", "file_provider_authority", "\"com.hedvig.android.test.file.provider\""
 
-            minifyEnabled true
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            // minifyEnabled true
+            // proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            debuggable true
         }
 
         debug {
@@ -90,6 +91,19 @@ android {
             debuggable true
         }
     }
+
+    sourceSets {
+        debug {
+            java.srcDirs += "src/engineering/java"
+            res.srcDirs += "src/engineering/res"
+            manifest.srcFile "src/debug/AndroidManifest.xml"
+        }
+        staging {
+            java.srcDirs += "src/engineering/java"
+            res.srcDirs += "src/engineering/res"
+            manifest.srcFile "src/debug/AndroidManifest.xml"
+        }
+    }
 }
 
 dependencies {
@@ -97,6 +111,7 @@ dependencies {
 
     androidTestImplementation project(':testdata')
     debugImplementation project(':testdata')
+    stagingImplementation project(':testdata')
 
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.0.10'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
