+++ app/src/debug/res/values/strings.xml
@@ -9,7 +9,6 @@
     <string name="alert_title" translatable="false">Title</string>
     <string name="alert_message" translatable="false">Message</string>
 
-    <string name="mock_personas_prompt" translatable="false">Mock Persona</string>
 
     <string name="facebook_app_id" translatable="false">437830083826480</string>
     <string name="fb_login_protocol_scheme" translatable="false">fb437830083826480</string>
