+++ app/src/engineering/java/com/hedvig/app/feature/chat/ChatMockActivity.kt
@@ -6,15 +6,17 @@ import androidx.core.os.bundleOf
 import com.google.firebase.messaging.RemoteMessage
 import com.hedvig.app.GenericDevelopmentAdapter
 import com.hedvig.app.R
+import com.hedvig.app.databinding.ActivityGenericDevelopmentBinding
 import com.hedvig.app.feature.chat.service.ChatNotificationManager
 import com.hedvig.app.feature.chat.service.ChatNotificationManager.DATA_NEW_MESSAGE_BODY
-import kotlinx.android.synthetic.debug.activity_generic_development.*
+import com.hedvig.app.util.extensions.viewBinding
 
 class ChatMockActivity : AppCompatActivity(R.layout.activity_generic_development) {
+    private val binding by viewBinding(ActivityGenericDevelopmentBinding::bind)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        root.adapter = GenericDevelopmentAdapter(
+        binding.root.adapter = GenericDevelopmentAdapter(
             listOf(
                 GenericDevelopmentAdapter.Item.Header("Notifications"),
                 GenericDevelopmentAdapter.Item.ClickableItem("Send chat message received-notification") {
