+++ app/src/engineering/res/layout/activity_vectory_drawable_gallery.xml
