+++ app/src/engineering/res/layout/activity_view_gallery.xml
