+++ app/src/engineering/res/layout/development_header.xml
@@ -28,7 +28,6 @@
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
         android:layout_marginTop="@dimen/base_margin"
-        android:entries="@array/mock_personas"
-        android:prompt="@string/mock_personas_prompt" />
+        android:entries="@array/mock_personas" />
 
 </LinearLayout>
