+++ app/src/engineering/res/layout/view_gallery_bottom_sheet.xml
